projects/agent-mcp/pyproject.toml
[project]
name = "agent-mcp"
version = "0.1.0"
description = "Job finder agent using OpenAI Agents SDK and Firecrawl MCP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents",
    "python-dotenv",
    "asyncio",
    "rich",
]
